Move cohorts to individual router
